L1-uGT emulator: fix value of 3-muon mass with "duplicate" muons
This change fixes the value assigned to the 3-muon mass
in the CorrThreeBodyCondition class used by the L1-uGT emulator
in the case where at least 2 of the 3 dimuon masses have identical values
(the latter case occurs in events containing multiple muons
with the same hwPt, hwEtaAtVtx, and hwPhiAtVtx).
It makes the logic used in the emulator
closer to that implemented in the L1-uGT firmware.

In the context of the Run-3 pp trigger menus,
this reduces (and possibly, removes) data-emulator mismatches
for L1T algorithms for tau-to-3-muons searches,
the only algorithms in those menus applying a cut on a 3-muon mass.